Make xbgpu pass mypy type-checking
Remove the line from mypy.ini that excludes xbgpu's code and tests from
checking. It all passes now.

Let mypy ignore the instances where we manually alter the dtype of an array.
There doesn't seem to be a way to get around this.

When altering the value of numpy-managed signed-integer variables, explicitly
create the right type. I'm not sure whether this is really the right approach
though, so I've left a note indicating as much.

Remove verify_reorder() from the precorrelation reorder test. It's superceded
by the ported-to-C function, and we can get it back from the git history if
we need to. It doesn't seem worth the effort (to me) to get it working if it's
not getting used anyway.

Change AbstractStream to _AsyncStream to reflect what is available in spead2's
stub files more accurately.

Change some inconsistent usages of cuda.Context as a type hint where something
marked as an AbstractContext is being passed in. You can give something
more specific as an argument when the general case is type-hinted, but not
the other way round. Also while I'm there, remove redundant type hints from
the docstring.

Resolves: NGC-256

+++ test/xbgpu/xbengine_test.py
@@ -119,14 +119,16 @@ def createHeaps(
 
             # 2.1.1 Make sure none of these samples are equal to -128 as that is not a supported value with the Tensor
             # cores.
+            # Note: using np.byte on the assignment satsifies mypy but there may
+            # be a more elegant way to do this.
             if pol0Real == -128:
-                pol0Real = -127
+                pol0Real = np.byte(-127)
             if pol0Imag == -128:
-                pol0Imag = -127
+                pol0Imag = np.byte(-127)
             if pol1Real == -128:
-                pol1Real = -127
+                pol1Real = np.byte(-127)
             if pol1Imag == -128:
-                pol1Imag = -127
+                pol1Imag = np.byte(-127)
 
             # 2.1.2 Combine values into a code word. The values are all cast to uint8s as when I was casting them to
             # int8s, the sign extension would behave strangly and what I expected to be in the code word would be
@@ -143,7 +145,8 @@ def createHeaps(
 
         # 2.2 Change dtype and shape of the array back to the correct values required by the receiver. The data itself
         # is not modified, its just how it is intepreted that is changed.
-        sample_array.dtype = np.int8
+        # Mypy will complain about this - the dtype is supposed to be read-only. But we're all consenting adults here.
+        sample_array.dtype = np.int8  # type: ignore
         sample_array = np.reshape(sample_array, heap_shape)
 
         # 2.3 Assign all values to the heap fields.
